+++ packages/gamut-system/src/core/system/__tests__/css-test.ts
@@ -16,4 +16,24 @@ describe('css', () => {
 
     expect(fn).toBeDefined();
   });
+
+  it('works with themes', () => {
+    const { css } = system.withTheme<{ fontSize: { sm: '14px' } }>().create({
+      typography: {
+        fontSize: { propName: 'fontSize', scale: 'fontSize' },
+      },
+    });
+
+    const fn = css({
+      '&:hover': {
+        fontSize: 'sm',
+      },
+      '&:before': {
+        content: '',
+        fontSize: 'sm',
+      },
+    });
+
+    expect(fn).toBeDefined();
+  });
 });
