+++ packages/gamut-system/src/types/config.ts
@@ -1,10 +1,10 @@
+import { CSSObject } from './css';
 import {
   DirectionalProperty,
   Properties,
   Property,
   PropName,
 } from './properties';
-import { CSSObject } from './system';
 import { SafeLookup, SafeMapKey, WeakRecord } from './utils';
 
 /** Theme Shape  */
